Splited status.h
